[ECO-2715] Update `BIGINT` to `NUMERIC` (#77)
* Change all `BIGINT` fields to `NUMERIC`

* Change `transaction_version`, `event_index`, and `block_number` to `BIGINT` fields to match the Aptos indexer processor column types

* Fix the aggregate_market_state function to count rows/events, too

* Update the schema with diesel migration

* Update all `i64` values in the json types to `BigDecimal` values

Update the `micros_to_naive_datetime` function to accept `BigDecimal`

* Fix test files

* Use `u64` where appropriate, convert and clone where necessary

+++ rust/processor/src/db/common/models/emojicoin_models/models/market_1m_periods_in_last_day.rs
@@ -18,8 +18,8 @@ use serde::{Deserialize, Serialize};
 #[diesel(primary_key(market_id, nonce))]
 #[diesel(table_name = market_1m_periods_in_last_day)]
 pub struct MarketOneMinutePeriodsInLastDayModel {
-    pub market_id: i64,
-    pub nonce: i64,
+    pub market_id: BigDecimal,
+    pub nonce: BigDecimal,
     pub transaction_version: i64,
     pub volume: BigDecimal,
     pub base_volume: BigDecimal,
