+++ rust/processor/src/db/postgres/migrations/2024-10-07-131051_fix_price_feed/up.sql
@@ -1,6 +1,6 @@
 -- Your SQL goes here
 CREATE OR REPLACE FUNCTION price_feed() RETURNS TABLE(
-  market_id BIGINT,
+  market_id NUMERIC,
   symbol_bytes BYTEA,
   symbol_emojis TEXT[],
   market_address VARCHAR(66),
